Initial commit
This repository is for deploying the paas-metric-exporter[1] as a Bosh
release. It needs to be able to run as a Bosh release because when
configured to send to Loggregator it will assume there is a Metron Agent
running on `localhost`; the easiest way to get Metron running is to
deploy it as a Bosh job colocated with existing Cloud Foundry
components.

[1] https://github.com/alphagov/paas-metric-exporter